+++ SBA_Serv/World/WorldMath.py
@@ -61,7 +61,7 @@ def aligninstances(obj1, obj2, class1, class2):
 def distancesquared(pos1x, pos1y, pos2x, pos2y):
     return (pos1x - pos2x) ** 2 + (pos1y - pos2y) ** 2
 
-def getPositionAwayFromOtherObjects(world, radius, edgebuffer, force=False):
+def getPositionAwayFromOtherObjects(world, radius, edgebuffer):
     """Returns a Position in the world that is away from other objects in it.
 
     Args:
@@ -76,7 +76,7 @@ def getPositionAwayFromOtherObjects(world, radius, edgebuffer, force=False):
     x = random.randint(edgebuffer, size[0] - edgebuffer)
     y = random.randint(edgebuffer, size[1] - edgebuffer)
     i = 0
-    while len(world.getObjectsInArea((x, y), radius, force)) > 0 and i < 15:
+    while len(world.getObjectsInArea((x, y), radius)) > 0 and i < 15:
         i += 1
         x = random.randint(edgebuffer, size[0] - edgebuffer)
         y = random.randint(edgebuffer, size[1] - edgebuffer)